+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to
 [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [v0.11.2] - 2021-08-25
 ### Fixed
 - Issue with `model_max_length` in tokenizer meant that models with an ill-set
   value of `max_position_embeddings` could not be benchmarked. Now, if